+++ js/gutenberg-availability.js
@@ -16,7 +16,7 @@ registerBlockType('recras/availability', {
 
     edit: withSelect((select) => {
         return {
-            packages: select('recras/store').fetchPackages(true),
+            packages: select('recras/store').fetchPackages(true, false),
             pagesPosts: select('recras/store').fetchPagesPosts(),
         }
     })(props => {