+++ js/gutenberg-voucher-sales.js
@@ -51,6 +51,11 @@ registerBlockType('recras/voucher-sales', {
             options: voucherTemplates,
             label: __('Voucher template', TEXT_DOMAIN),
         };
+        if (voucherTemplates.length === 1) {
+            props.setAttributes({
+                id: voucherTemplates[0].value,
+            });
+        }
 
         const optionsRedirectControl = {
             value: redirect,