+++ changelog.md
@@ -3,6 +3,7 @@
 ## unreleased
 * Fix whitespace in online booking/voucher sales causing problems in edge cases
 * Update online booking library version: this fixes 'require X per Y' requirements (instead of just 1 per Y)
+* Fix "Error: no ID set" when only having a single contact form/package/product/voucher template in Gutenberg blocks
 
 ## 3.1.0 (2019-11-12)
 * Improve online booking styling in Internet Explorer